Rename to "collision_kernels"
