+++ PySDM_tests/smoke_tests/Shima_et_al_2009/test_coalescence.py
@@ -4,7 +4,7 @@ import pytest
 from PySDM.backends import ThrustRTC
 from PySDM.builder import Builder
 from PySDM.dynamics import Coalescence
-from PySDM.physics.coalescence_kernels import Golovin
+from PySDM.physics.collision_kernels import Golovin
 from PySDM.environments import Box
 from PySDM.initialisation.spectra import Exponential
 from PySDM.initialisation.spectral_sampling import ConstantMultiplicity
